Allow downcasts of Literals if conversion is lossless
The patch makes it possible to downcast Literals with higher precedence in expressions which also contain other castable arguments.

```
Before: int_arr[1] = 123 => to_long(subscript(int_arr, 1)) = 123
          ^^^^^      ^^^
         castable castable
After:  int_arr[1] = 123 => subscript(int_arr, 1) = to_int(123)
```

This has the advantage that the lossless conversion is only performed once for
the Literal, whereas it might have to be performed for every record in case of a
column. The LuceneQueryBuilder is only able to generate a fast Lucene index
query if no cast function is applied.